users_api/requirements.txt
Django==4.0.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
